# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_minisource_log_internal_models.LogAlert:
    properties:
      channels:
        items:
          type: integer
        type: array
      created_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      filter:
        items:
          type: integer
        type: array
      id:
        type: string
      last_triggered:
        type: string
      name:
        type: string
      severity:
        type: string
      tenant_id:
        type: string
      threshold:
        type: integer
      updated_at:
        type: string
      window_mins:
        type: integer
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:5002
info:
  contact: {}
  description: Centralized logging service for Minisource
  title: Log Service API
  version: "1.0"
paths:
  /alerts:
    get:
      description: Lists all alerts for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_minisource_log_internal_models.LogAlert'
            type: array
      summary: List alerts
      tags:
      - alerts
    post:
      consumes:
      - application/json
      description: Creates a new log alert rule
      parameters:
      - description: Log Alert
        in: body
        name: alert
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_log_internal_models.LogAlert'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_minisource_log_internal_models.LogAlert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Create alert
      tags:
      - alerts
  /alerts/{id}:
    delete:
      description: Deletes an alert
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Delete alert
      tags:
      - alerts
    get:
      description: Retrieves an alert by ID
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_log_internal_models.LogAlert'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get alert
      tags:
      - alerts
    put:
      consumes:
      - application/json
      description: Updates an existing alert
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Log Alert
        in: body
        name: alert
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_log_internal_models.LogAlert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_log_internal_models.LogAlert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Update alert
      tags:
      - alerts
  /alerts/{id}/disable:
    post:
      description: Disables an alert
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Disable alert
      tags:
      - alerts
  /alerts/{id}/enable:
    post:
      description: Enables an alert
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Enable alert
      tags:
      - alerts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"